fix: tooling library not functioning
