+++ __tests__/tooling/lib/find-schema-definition.test.js
@@ -1,5 +1,5 @@
 const petstore = require('@readme/oas-examples/3.0/json/petstore');
-const findSchemaDefinition = require('../../src/lib/find-schema-definition');
+const findSchemaDefinition = require('../../../tooling/lib/find-schema-definition');
 
 test('should return a definition for a given ref', () => {
   expect(findSchemaDefinition('#/components/schemas/Pet', petstore)).toStrictEqual(petstore.components.schemas.Pet);
