+++ __tests__/tooling/lib/flatten-array.test.js
@@ -1,4 +1,4 @@
-const flattenArray = require('../../src/lib/flatten-array');
+const flattenArray = require('../../../tooling/lib/flatten-array');
 
 test('should flatten array', () => {
   const array = [[1], [2, 3], [[4, 5]]];
