+++ package.json
@@ -58,6 +58,7 @@
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@readme/eslint-config": "^3.6.2",
+    "@readme/oas-examples": "^4.0.0",
     "conventional-changelog-cli": "^2.1.0",
     "eslint": "^7.11.0",
     "husky": "^4.3.0",
@@ -69,10 +70,11 @@
   "prettier": "@readme/eslint-config/prettier",
   "jest": {
     "coveragePathIgnorePatterns": [
-      "__tests__/**/__fixtures__/"
+      "<rootDir>/node_modules",
+      "<rootDir>/__tests__/(.*)/__fixtures__/"
     ],
     "testPathIgnorePatterns": [
-      "__tests__/**/__fixtures__/"
+      "<rootDir>/__tests__/(.*)/__fixtures__/"
     ]
   },
   "husky": {
