+++ tooling/lib/flatten-array.js
