+++ tooling/lib/flatten-schema.js
