+++ tooling/lib/get-path-operation.js
